Add files via upload
This lab exam your android design skills will be evaluated. For this you need to use Android XML. 
This project aims to assess your ability to create functional, aesthetically pleasing, and user-friendly UI designs that adhere to Android development best practices.
We are tasked with designing the UI for an app idea that we are passionate about or find 
interesting. This can be anything from a productivity app, a social media platform, a fitness 
tracker, to a game. The key is to showcase your creativity, understanding of UI/UX principles, 
and your ability to implement these designs using XML in Android Studio.