Add files via upload
I tasked with designing the UI for an app idea that I am passionate about or find 
interesting. This can be anything from a productivity app, a social media platform, a fitness 
tracker, to a game. The key is to showcase your creativity, understanding of UI/UX principles, 
and your ability to implement these designs using XML in Android Studio.